Revert "Filter interface change (#102)"
This reverts commit b6c5446319fee6343aa6111d52affbe5cf483317.

+++ test/foundry/lib/ds-test/test.sol
@@ -16,37 +16,33 @@
 pragma solidity >=0.5.0;
 
 contract DSTest {
-    event log(string);
-    event logs(bytes);
+    event log                    (string);
+    event logs                   (bytes);
 
-    event log_address(address);
-    event log_bytes32(bytes32);
-    event log_int(int256);
-    event log_uint(uint256);
-    event log_bytes(bytes);
-    event log_string(string);
+    event log_address            (address);
+    event log_bytes32            (bytes32);
+    event log_int                (int);
+    event log_uint               (uint);
+    event log_bytes              (bytes);
+    event log_string             (string);
 
-    event log_named_address(string key, address val);
-    event log_named_bytes32(string key, bytes32 val);
-    event log_named_decimal_int(string key, int256 val, uint256 decimals);
-    event log_named_decimal_uint(string key, uint256 val, uint256 decimals);
-    event log_named_int(string key, int256 val);
-    event log_named_uint(string key, uint256 val);
-    event log_named_bytes(string key, bytes val);
-    event log_named_string(string key, string val);
+    event log_named_address      (string key, address val);
+    event log_named_bytes32      (string key, bytes32 val);
+    event log_named_decimal_int  (string key, int val, uint decimals);
+    event log_named_decimal_uint (string key, uint val, uint decimals);
+    event log_named_int          (string key, int val);
+    event log_named_uint         (string key, uint val);
+    event log_named_bytes        (string key, bytes val);
+    event log_named_string       (string key, string val);
 
     bool public IS_TEST = true;
     bool private _failed;
 
-    address constant HEVM_ADDRESS = address(bytes20(uint160(uint256(keccak256("hevm cheat code")))));
+    address constant HEVM_ADDRESS =
+    address(bytes20(uint160(uint256(keccak256('hevm cheat code')))));
 
-    modifier mayRevert() {
-        _;
-    }
-
-    modifier testopts(string memory) {
-        _;
-    }
+    modifier mayRevert() { _; }
+    modifier testopts(string memory) { _; }
 
     function failed() public returns (bool) {
         if (_failed) {
@@ -56,7 +52,8 @@ contract DSTest {
             if (hasHEVMContext()) {
                 (, bytes memory retdata) = HEVM_ADDRESS.call(
                     abi.encodePacked(
-                        bytes4(keccak256("load(address,bytes32)")), abi.encode(HEVM_ADDRESS, bytes32("failed"))
+                        bytes4(keccak256("load(address,bytes32)")),
+                        abi.encode(HEVM_ADDRESS, bytes32("failed"))
                     )
                 );
                 globalFailed = abi.decode(retdata, (bool));
@@ -67,7 +64,7 @@ contract DSTest {
 
     function fail() internal {
         if (hasHEVMContext()) {
-            (bool status,) = HEVM_ADDRESS.call(
+            (bool status, ) = HEVM_ADDRESS.call(
                 abi.encodePacked(
                     bytes4(keccak256("store(address,bytes32,bytes32)")),
                     abi.encode(HEVM_ADDRESS, bytes32("failed"), bytes32(uint256(0x01)))
@@ -87,9 +84,9 @@ contract DSTest {
     }
 
     modifier logs_gas() {
-        uint256 startGas = gasleft();
+        uint startGas = gasleft();
         _;
-        uint256 endGas = gasleft();
+        uint endGas = gasleft();
         emit log_named_uint("gas", startGas - endGas);
     }
 
@@ -115,10 +112,9 @@ contract DSTest {
             fail();
         }
     }
-
     function assertEq(address a, address b, string memory err) internal {
         if (a != b) {
-            emit log_named_string("Error", err);
+            emit log_named_string ("Error", err);
             assertEq(a, b);
         }
     }
@@ -131,23 +127,20 @@ contract DSTest {
             fail();
         }
     }
-
     function assertEq(bytes32 a, bytes32 b, string memory err) internal {
         if (a != b) {
-            emit log_named_string("Error", err);
+            emit log_named_string ("Error", err);
             assertEq(a, b);
         }
     }
-
     function assertEq32(bytes32 a, bytes32 b) internal {
         assertEq(a, b);
     }
-
     function assertEq32(bytes32 a, bytes32 b, string memory err) internal {
         assertEq(a, b, err);
     }
 
-    function assertEq(int256 a, int256 b) internal {
+    function assertEq(int a, int b) internal {
         if (a != b) {
             emit log("Error: a == b not satisfied [int]");
             emit log_named_int("  Expected", b);
@@ -155,15 +148,13 @@ contract DSTest {
             fail();
         }
     }
-
-    function assertEq(int256 a, int256 b, string memory err) internal {
+    function assertEq(int a, int b, string memory err) internal {
         if (a != b) {
             emit log_named_string("Error", err);
             assertEq(a, b);
         }
     }
-
-    function assertEq(uint256 a, uint256 b) internal {
+    function assertEq(uint a, uint b) internal {
         if (a != b) {
             emit log("Error: a == b not satisfied [uint]");
             emit log_named_uint("  Expected", b);
@@ -171,15 +162,13 @@ contract DSTest {
             fail();
         }
     }
-
-    function assertEq(uint256 a, uint256 b, string memory err) internal {
+    function assertEq(uint a, uint b, string memory err) internal {
         if (a != b) {
             emit log_named_string("Error", err);
             assertEq(a, b);
         }
     }
-
-    function assertEqDecimal(int256 a, int256 b, uint256 decimals) internal {
+    function assertEqDecimal(int a, int b, uint decimals) internal {
         if (a != b) {
             emit log("Error: a == b not satisfied [decimal int]");
             emit log_named_decimal_int("  Expected", b, decimals);
@@ -187,15 +176,13 @@ contract DSTest {
             fail();
         }
     }
-
-    function assertEqDecimal(int256 a, int256 b, uint256 decimals, string memory err) internal {
+    function assertEqDecimal(int a, int b, uint decimals, string memory err) internal {
         if (a != b) {
             emit log_named_string("Error", err);
             assertEqDecimal(a, b, decimals);
         }
     }
-
-    function assertEqDecimal(uint256 a, uint256 b, uint256 decimals) internal {
+    function assertEqDecimal(uint a, uint b, uint decimals) internal {
         if (a != b) {
             emit log("Error: a == b not satisfied [decimal uint]");
             emit log_named_decimal_uint("  Expected", b, decimals);
@@ -203,15 +190,14 @@ contract DSTest {
             fail();
         }
     }
-
-    function assertEqDecimal(uint256 a, uint256 b, uint256 decimals, string memory err) internal {
+    function assertEqDecimal(uint a, uint b, uint decimals, string memory err) internal {
         if (a != b) {
             emit log_named_string("Error", err);
             assertEqDecimal(a, b, decimals);
         }
     }
 
-    function assertGt(uint256 a, uint256 b) internal {
+    function assertGt(uint a, uint b) internal {
         if (a <= b) {
             emit log("Error: a > b not satisfied [uint]");
             emit log_named_uint("  Value a", a);
@@ -219,15 +205,13 @@ contract DSTest {
             fail();
         }
     }
-
-    function assertGt(uint256 a, uint256 b, string memory err) internal {
+    function assertGt(uint a, uint b, string memory err) internal {
         if (a <= b) {
             emit log_named_string("Error", err);
             assertGt(a, b);
         }
     }
-
-    function assertGt(int256 a, int256 b) internal {
+    function assertGt(int a, int b) internal {
         if (a <= b) {
             emit log("Error: a > b not satisfied [int]");
             emit log_named_int("  Value a", a);
@@ -235,15 +219,13 @@ contract DSTest {
             fail();
         }
     }
-
-    function assertGt(int256 a, int256 b, string memory err) internal {
+    function assertGt(int a, int b, string memory err) internal {
         if (a <= b) {
             emit log_named_string("Error", err);
             assertGt(a, b);
         }
     }
-
-    function assertGtDecimal(int256 a, int256 b, uint256 decimals) internal {
+    function assertGtDecimal(int a, int b, uint decimals) internal {
         if (a <= b) {
             emit log("Error: a > b not satisfied [decimal int]");
             emit log_named_decimal_int("  Value a", a, decimals);
@@ -251,15 +233,13 @@ contract DSTest {
             fail();
         }
     }
-
-    function assertGtDecimal(int256 a, int256 b, uint256 decimals, string memory err) internal {
+    function assertGtDecimal(int a, int b, uint decimals, string memory err) internal {
         if (a <= b) {
             emit log_named_string("Error", err);
             assertGtDecimal(a, b, decimals);
         }
     }
-
-    function assertGtDecimal(uint256 a, uint256 b, uint256 decimals) internal {
+    function assertGtDecimal(uint a, uint b, uint decimals) internal {
         if (a <= b) {
             emit log("Error: a > b not satisfied [decimal uint]");
             emit log_named_decimal_uint("  Value a", a, decimals);
@@ -267,15 +247,14 @@ contract DSTest {
             fail();
         }
     }
-
-    function assertGtDecimal(uint256 a, uint256 b, uint256 decimals, string memory err) internal {
+    function assertGtDecimal(uint a, uint b, uint decimals, string memory err) internal {
         if (a <= b) {
             emit log_named_string("Error", err);
             assertGtDecimal(a, b, decimals);
         }
     }
 
-    function assertGe(uint256 a, uint256 b) internal {
+    function assertGe(uint a, uint b) internal {
         if (a < b) {
             emit log("Error: a >= b not satisfied [uint]");
             emit log_named_uint("  Value a", a);
@@ -283,15 +262,13 @@ contract DSTest {
             fail();
         }
     }
-
-    function assertGe(uint256 a, uint256 b, string memory err) internal {
+    function assertGe(uint a, uint b, string memory err) internal {
         if (a < b) {
             emit log_named_string("Error", err);
             assertGe(a, b);
         }
     }
-
-    function assertGe(int256 a, int256 b) internal {
+    function assertGe(int a, int b) internal {
         if (a < b) {
             emit log("Error: a >= b not satisfied [int]");
             emit log_named_int("  Value a", a);
@@ -299,15 +276,13 @@ contract DSTest {
             fail();
         }
     }
-
-    function assertGe(int256 a, int256 b, string memory err) internal {
+    function assertGe(int a, int b, string memory err) internal {
         if (a < b) {
             emit log_named_string("Error", err);
             assertGe(a, b);
         }
     }
-
-    function assertGeDecimal(int256 a, int256 b, uint256 decimals) internal {
+    function assertGeDecimal(int a, int b, uint decimals) internal {
         if (a < b) {
             emit log("Error: a >= b not satisfied [decimal int]");
             emit log_named_decimal_int("  Value a", a, decimals);
@@ -315,15 +290,13 @@ contract DSTest {
             fail();
         }
     }
-
-    function assertGeDecimal(int256 a, int256 b, uint256 decimals, string memory err) internal {
+    function assertGeDecimal(int a, int b, uint decimals, string memory err) internal {
         if (a < b) {
             emit log_named_string("Error", err);
             assertGeDecimal(a, b, decimals);
         }
     }
-
-    function assertGeDecimal(uint256 a, uint256 b, uint256 decimals) internal {
+    function assertGeDecimal(uint a, uint b, uint decimals) internal {
         if (a < b) {
             emit log("Error: a >= b not satisfied [decimal uint]");
             emit log_named_decimal_uint("  Value a", a, decimals);
@@ -331,15 +304,14 @@ contract DSTest {
             fail();
         }
     }
-
-    function assertGeDecimal(uint256 a, uint256 b, uint256 decimals, string memory err) internal {
+    function assertGeDecimal(uint a, uint b, uint decimals, string memory err) internal {
         if (a < b) {
             emit log_named_string("Error", err);
             assertGeDecimal(a, b, decimals);
         }
     }
 
-    function assertLt(uint256 a, uint256 b) internal {
+    function assertLt(uint a, uint b) internal {
         if (a >= b) {
             emit log("Error: a < b not satisfied [uint]");
             emit log_named_uint("  Value a", a);
@@ -347,15 +319,13 @@ contract DSTest {
             fail();
         }
     }
-
-    function assertLt(uint256 a, uint256 b, string memory err) internal {
+    function assertLt(uint a, uint b, string memory err) internal {
         if (a >= b) {
             emit log_named_string("Error", err);
             assertLt(a, b);
         }
     }
-
-    function assertLt(int256 a, int256 b) internal {
+    function assertLt(int a, int b) internal {
         if (a >= b) {
             emit log("Error: a < b not satisfied [int]");
             emit log_named_int("  Value a", a);
@@ -363,15 +333,13 @@ contract DSTest {
             fail();
         }
     }
-
-    function assertLt(int256 a, int256 b, string memory err) internal {
+    function assertLt(int a, int b, string memory err) internal {
         if (a >= b) {
             emit log_named_string("Error", err);
             assertLt(a, b);
         }
     }
-
-    function assertLtDecimal(int256 a, int256 b, uint256 decimals) internal {
+    function assertLtDecimal(int a, int b, uint decimals) internal {
         if (a >= b) {
             emit log("Error: a < b not satisfied [decimal int]");
             emit log_named_decimal_int("  Value a", a, decimals);
@@ -379,15 +347,13 @@ contract DSTest {
             fail();
         }
     }
-
-    function assertLtDecimal(int256 a, int256 b, uint256 decimals, string memory err) internal {
+    function assertLtDecimal(int a, int b, uint decimals, string memory err) internal {
         if (a >= b) {
             emit log_named_string("Error", err);
             assertLtDecimal(a, b, decimals);
         }
     }
-
-    function assertLtDecimal(uint256 a, uint256 b, uint256 decimals) internal {
+    function assertLtDecimal(uint a, uint b, uint decimals) internal {
         if (a >= b) {
             emit log("Error: a < b not satisfied [decimal uint]");
             emit log_named_decimal_uint("  Value a", a, decimals);
@@ -395,15 +361,14 @@ contract DSTest {
             fail();
         }
     }
-
-    function assertLtDecimal(uint256 a, uint256 b, uint256 decimals, string memory err) internal {
+    function assertLtDecimal(uint a, uint b, uint decimals, string memory err) internal {
         if (a >= b) {
             emit log_named_string("Error", err);
             assertLtDecimal(a, b, decimals);
         }
     }
 
-    function assertLe(uint256 a, uint256 b) internal {
+    function assertLe(uint a, uint b) internal {
         if (a > b) {
             emit log("Error: a <= b not satisfied [uint]");
             emit log_named_uint("  Value a", a);
@@ -411,15 +376,13 @@ contract DSTest {
             fail();
         }
     }
-
-    function assertLe(uint256 a, uint256 b, string memory err) internal {
+    function assertLe(uint a, uint b, string memory err) internal {
         if (a > b) {
             emit log_named_string("Error", err);
             assertLe(a, b);
         }
     }
-
-    function assertLe(int256 a, int256 b) internal {
+    function assertLe(int a, int b) internal {
         if (a > b) {
             emit log("Error: a <= b not satisfied [int]");
             emit log_named_int("  Value a", a);
@@ -427,15 +390,13 @@ contract DSTest {
             fail();
         }
     }
-
-    function assertLe(int256 a, int256 b, string memory err) internal {
+    function assertLe(int a, int b, string memory err) internal {
         if (a > b) {
             emit log_named_string("Error", err);
             assertLe(a, b);
         }
     }
-
-    function assertLeDecimal(int256 a, int256 b, uint256 decimals) internal {
+    function assertLeDecimal(int a, int b, uint decimals) internal {
         if (a > b) {
             emit log("Error: a <= b not satisfied [decimal int]");
             emit log_named_decimal_int("  Value a", a, decimals);
@@ -443,15 +404,13 @@ contract DSTest {
             fail();
         }
     }
-
-    function assertLeDecimal(int256 a, int256 b, uint256 decimals, string memory err) internal {
+    function assertLeDecimal(int a, int b, uint decimals, string memory err) internal {
         if (a > b) {
             emit log_named_string("Error", err);
             assertLeDecimal(a, b, decimals);
         }
     }
-
-    function assertLeDecimal(uint256 a, uint256 b, uint256 decimals) internal {
+    function assertLeDecimal(uint a, uint b, uint decimals) internal {
         if (a > b) {
             emit log("Error: a <= b not satisfied [decimal uint]");
             emit log_named_decimal_uint("  Value a", a, decimals);
@@ -459,8 +418,7 @@ contract DSTest {
             fail();
         }
     }
-
-    function assertLeDecimal(uint256 a, uint256 b, uint256 decimals, string memory err) internal {
+    function assertLeDecimal(uint a, uint b, uint decimals, string memory err) internal {
         if (a > b) {
             emit log_named_string("Error", err);
             assertGeDecimal(a, b, decimals);
@@ -475,7 +433,6 @@ contract DSTest {
             fail();
         }
     }
-
     function assertEq(string memory a, string memory b, string memory err) internal {
         if (keccak256(abi.encodePacked(a)) != keccak256(abi.encodePacked(b))) {
             emit log_named_string("Error", err);
@@ -486,7 +443,7 @@ contract DSTest {
     function checkEq0(bytes memory a, bytes memory b) internal pure returns (bool ok) {
         ok = true;
         if (a.length == b.length) {
-            for (uint256 i = 0; i < a.length; i++) {
+            for (uint i = 0; i < a.length; i++) {
                 if (a[i] != b[i]) {
                     ok = false;
                 }
@@ -495,7 +452,6 @@ contract DSTest {
             ok = false;
         }
     }
-
     function assertEq0(bytes memory a, bytes memory b) internal {
         if (!checkEq0(a, b)) {
             emit log("Error: a == b not satisfied [bytes]");
@@ -504,7 +460,6 @@ contract DSTest {
             fail();
         }
     }
-
     function assertEq0(bytes memory a, bytes memory b, string memory err) internal {
         if (!checkEq0(a, b)) {
             emit log_named_string("Error", err);
